refactor: add service interfaces
