+++ src/main/java/com/aegro/api/service/FarmService.java
@@ -3,31 +3,16 @@ package com.aegro.api.service;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.aegro.api.entities.Farm;
-import com.aegro.api.repository.FarmRepository;
 
-@Service
-public class FarmService {
-
-	@Autowired
-	private FarmRepository farmRepository;
+public interface FarmService {
 	
-	public Farm saveFarm(Farm farm) {
-		return farmRepository.save(farm); 
-	}
-
-	public List<Farm> farmList() {
-		return farmRepository.findAll();
-	}
+	public Farm saveFarm(Farm farm);
 	
-	public Optional<Farm> getFarmById(Long id){
-		return farmRepository.findById(id);
-	}
+	public List<Farm> farmList();
 	
-	public void removeFarmById(Long id) {
-		farmRepository.deleteById(id);	
-	}
+	public Optional<Farm> getFarmById(Long id);
+	
+	public void removeFarmById(Long id);
+	
 }
