+++ src/main/java/com/aegro/api/service/ProductionService.java
@@ -3,36 +3,19 @@ package com.aegro.api.service;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.aegro.api.entities.Farm;
 import com.aegro.api.entities.Production;
-import com.aegro.api.repository.ProductionRepository;
 
-@Service
-public class ProductionService {
-	
-	@Autowired
-	private ProductionRepository productionRepository;
-	
-	public Production saveProduction(Production production) {
-		return productionRepository.save(production); 
-	}
 
-	public List<Production> productionList() {
-		return productionRepository.findAll();
-	}
+public interface ProductionService {
+
+	public Production saveProduction(Production production);
+
+	public List<Production> productionList();
 	
-	public Optional<Production> getProductionById(Long id){
-		return productionRepository.findById(id);
-	}
+	public Optional<Production> getProductionById(Long id);
 	
-	public void removeProductionById(Long id) {
-		productionRepository.deleteById(id);	
-	}
+	public void removeProductionById(Long id);
 	
-	public Short totalProductionPerPlot (Long idPlot) {
-		return productionRepository.totalProductionByPlot(idPlot);
-	}
+	public Integer totalProductionPerPlot (Long idPlot);
 }
+
